finshed group endpoints and testing

+++ server/routes/groupRoutes.js
@@ -1,5 +1,6 @@
 const express = require('express');
 const { getGroup, getGroupUsers, getGroupExpenses, addGroup, addGroupUser, deleteGroup, deleteGroupUser } = require('../models/groupModel');
+const { getUser, getUserGroups } = require('../models/userModel');
 const { withErrorHandling } = require('./utils');
 
 const router = express.Router();
@@ -10,7 +11,7 @@ router.get('/:id', withErrorHandling(async (req, res) => {
     const groupID = req.params.id;
     const group = await getGroup(groupID);
     if (!group) {
-        return res.status(400).send('Group not found');
+        return res.status(404).send('Group not found');
     }
     return res.json(group);
 }));
@@ -40,30 +41,50 @@ router.get('/:id/expenses', withErrorHandling(async (req, res) => {
 // Contact through firebase??
 router.post('/', withErrorHandling(async (req, res) => {
     const { groupName, initialMembers } = req.body
+    if (!groupName) {
+        return res.status(400).send('Must supply group name');
+    }
     // TODO deal with emailing initialMembers
     const newGroup = await addGroup(groupName);
     return res.status(200).send('OK')
 }));
 
-router.post(':/id/addUser/:userID', withErrorHandling(async (req, res) => {
-    const { groupID, userID } = req.params.id
-    const userGroup = await addGroupUser(groupID, userID);
-    if (!userGroup) {
-        console.log('User or group specified by IDs does not exist')
+router.post('/:id/addUser', withErrorHandling(async (req, res) => {
+    const groupID = req.params.id;
+    const { userID } = req.body;
+    const userCheck = await getUser(userID);
+    const groupCheck = await getGroup(groupID);
+    if (!userCheck || !groupCheck) {
+        return res.status(404).send('User or group specified by ID does not exist');
     }
-    res.json(userGroup)
+    const userGroups = await getUserGroups(userID);
+    const duplicateCheck = userGroups.find(group => group.groupID == groupID);
+    if (duplicateCheck) {
+        return res.status(409).send(`User is already a member of that group`);
+    }
+    const userGroup = await addGroupUser(groupID, userID);
+    return res.status(200).send('OK');
 }));
 
 // DELETE ROUTES
 
-router.delete(':/id', withErrorHandling(async (req, res) => {
+router.delete('/:id', withErrorHandling(async (req, res) => {
     const groupID = req.params.id;
+    if (!getGroup(groupID)) {
+        return res.status(400).send('Group does not exist');
+    }
     const result = await deleteGroup(groupID);
     return res.status(200).send('OK');
 }));
 
 router.delete('/:groupID/removeUser/:userID', withErrorHandling(async (req, res) => {
     const { groupID, userID } = req.params;
+    if (!getGroup(groupID)) {
+        return res.status(400).send('Group does not exist');
+    }
+    if (!getUser(userID)) {
+        return res.status(400).send('User is not a member of that group');
+    }
     const result = await deleteGroupUser(groupID, userID);
     return res.status(200).send('OK');
 }));
